+++ R/ms_synthesize.R
@@ -6,8 +6,8 @@
 #' provided.
 #' @param script A character vector of lines to be spoken
 #' @param token An authentication token, base-64 encoded usually from
-#' \code{\link{get_ms_tts_token}}.  If not provided, will be created from
-#' \code{\link{get_ms_tts_token}}
+#' \code{\link{ms_get_tts_token}}.  If not provided, will be created from
+#' \code{\link{ms_get_tts_token}}
 #' @param gender Sex of the Speaker
 #' @param language Language to be spoken,
 #' must be from \code{\link{ms_language_codes}}
@@ -15,7 +15,10 @@
 #' \url{https://docs.microsoft.com/en-us/azure/cognitive-services/speech/api-reference-rest/bingvoiceoutput}
 #' for more information
 #' @param escape Should non-standard characters be substituted?  Should not
-#' be used if \code{script} has SSML tags. See \code{\link{create_ssml}}
+#' be used if \code{script} has SSML tags. See \code{\link{ms_create_ssml}}
+#' @param voice full voice name, usually from
+#' \code{\link{ms_language_to_ms_name}}.  Will override
+#' language and gender.
 #' @param ... Additional arguments to send to \code{\link{POST}}
 #'
 #' @return A list of the request, content, token, and `SSML`.
@@ -25,11 +28,11 @@
 #' then see below example
 #'
 #' @examples \dontrun{
-#' if (have_ms_tts_key()) {
+#' if (ms_have_tts_key()) {
 #' res = ms_synthesize(
 #' script = "hey, how are you doing? I'm doing pretty good",
 #' output_format = "audio-16khz-128kbitrate-mono-mp3")
-#' tmp <- tempfile()
+#' tmp <- tempfile(fileext = ".mp3")
 #' writeBin(res$content, con = tmp)
 #' mp3 = tuneR::readMP3(tmp)
 #' }
@@ -43,39 +46,38 @@ ms_synthesize = function(
   api_key = NULL,
   gender = c("Female", "Male"),
   language = "en-US",
-  output_format = c( "raw-16khz-16bit-mono-pcm" ,
-                     "ssml-16khz-16bit-mono-tts",
-                     "audio-16khz-16kbps-mono-siren",
-                     "riff-16khz-16kbps-mono-siren",
-                     "riff-16khz-16bit-mono-pcm",
-                     "audio-16khz-128kbitrate-mono-mp3",
-                     "audio-16khz-64kbitrate-mono-mp3",
-                     "audio-16khz-32kbitrate-mono-mp3"),
+  voice = NULL,
+  output_format = c("raw-16khz-16bit-mono-pcm", "raw-8khz-8bit-mono-mulaw",
+                    "riff-8khz-8bit-mono-alaw", "riff-8khz-8bit-mono-mulaw",
+                    "riff-16khz-16bit-mono-pcm", "audio-16khz-128kbitrate-mono-mp3",
+                    "audio-16khz-64kbitrate-mono-mp3", "audio-16khz-32kbitrate-mono-mp3",
+                    "raw-24khz-16bit-mono-pcm", "riff-24khz-16bit-mono-pcm",
+                    "audio-24khz-160kbitrate-mono-mp3", "audio-24khz-96kbitrate-mono-mp3",
+                    "audio-24khz-48kbitrate-mono-mp3"),
   escape = FALSE,
   region = NULL,
+  api = c("tts", "bing"),
   ...
 ){
 
-  # language_to_ms_name(langu)
-
-  L = validate_language_gender(
-    language = language,
-    gender = gender)
+  if (!is.null(voice)) {
+    L = ms_voice_info(voice)
+  } else {
+    L = ms_validate_language_gender(
+      language = language,
+      gender = gender)
+  }
   language = L$language
   gender = L$gender
-  xname = L$full_name
+  xname = L$full_name[1]
 
-
-  # synth_url = paste0(
-  #   'https://speech.platform.bing.com/',
-  #   'synthesize')
-  region = ms_region(region)
-  synth_url = paste0("https://", region,
-                     ".tts.speech.microsoft.com/",
-                     "cognitiveservices/v1")
+  synth_url = ms_synthesize_api_url(
+    api = api,
+    region = region
+  )
 
   if (is.null(token)) {
-    token = get_ms_tts_token(api_key = api_key)$token
+    token = ms_get_tts_token(api_key = api_key)$token
   }
 
   auth_hdr = add_headers(
@@ -87,8 +89,12 @@ ms_synthesize = function(
 
   ctype = content_type("application/ssml+xml")
 
-  ssml = create_ssml(script = script, gender = gender, language = language,
-                     escape = escape)
+  ssml = ms_create_ssml(
+    script = script,
+    gender = gender,
+    language = language,
+    voice = voice,
+    escape = escape)
 
   if (nchar(ssml) > 1024) {
     cat(ssml)
@@ -106,16 +112,100 @@ ms_synthesize = function(
     request = res,
     ssml = ssml,
     content = out,
-    token = token)
+    token = token,
+    output_format = output_format
+  )
   return(L)
-
 }
+
+#' Read Synthesized output
+#'
+#' @param output List from \code{\link{ms_synthesize}} with elements
+#' \code{output_format} and \code{content}
+#'
+#' @note The \code{tuneR} package cannot read all different types of
+#' the output here.
+#'
+#' @return A Wave Object
+#' @export
+ms_read_synthesis = function(output) {
+  tmp = tempfile()
+  writeBin(output$content, con = tmp)
+  output_format = output$output_format
+  wav = grepl("riff", tolower(output_format))
+  mp3 = grepl("mp3", tolower(output_format))
+  if (!wav & !mp3) {
+    warning("No format determined, assuming it's a WAV")
+    wav = TRUE
+  }
+
+  if (wav) {
+    out = tuneR::readWave(tmp)
+  }
+  if (mp3) {
+    out = tuneR::readMP3(tmp)
+  }
+  return(out)
+}
+
 
 #' @rdname ms_synthesize
 #' @param region Subscription region for your key.
 #' See \url{https://docs.microsoft.com/en-us/azure/cognitive-services/speech-service/rest-apis#text-to-speech}
 #'
 #' @export
-ms_region = function(region = c("westus", "eastasia", "northeurope")) {
+ms_region = function(region = ms_regions()) {
+  if (missing(region)) {
+    region = getOption("ms_region")
+  }
   region = match.arg(region)
+  return(region)
+}
+
+#' @rdname ms_synthesize
+#' @export
+#' @examples
+#' ms_regions()
+ms_regions = function() {
+  c("westus", "eastasia", "northeurope",
+    "westus2",
+    "eastus",
+    "eastus2",
+    "southeastasia",
+    "westeurope")
+}
+
+#' @rdname ms_synthesize
+#' @export
+ms_set_region = function(
+  region = ms_regions()) {
+  region = match.arg(region)
+  options(ms_region = region)
+  return(region)
+}
+
+
+
+#' @rdname ms_synthesize
+#' @param api which API to authorize on, either
+#' \code{tts} for text to speech or \code{bing} for
+#' Bing text to speech API
+#' @export
+ms_synthesize_api_url = function(
+  api = c("tts", "bing"),
+  region = NULL
+){
+  api = match.arg(api)
+
+  region = ms_region(region)
+  synth_url = switch(
+    api,
+    bing = paste0(
+      'https://speech.platform.bing.com/',
+      'synthesize'),
+    tts = paste0("https://", region,
+                 ".tts.speech.microsoft.com/",
+                 "cognitiveservices/v1")
+  )
+  return(synth_url)
 }