+++ NEWS.md
@@ -1,3 +1,13 @@
+# mscstts 0.5.1
+
+* Added voice listing.
+* Added new output formats for audio that should be correct.
+
+# mscstts 0.5.0
+
+* Added `ms_regions` to allow for the region for tokens.
+* Now supports TTS proper, not really Bing. 
+
 # mscstts 0.4.0
 
 * Added `reveal` code for tokens.